Isolate signal test to UNIX only
